product

+++ src/main/java/com/jun/portalservice/domain/services/VoucherService.java
@@ -1,11 +1,15 @@
 package com.jun.portalservice.domain.services;
 
+import com.jun.portalservice.app.dtos.VoucherDTO;
+import com.jun.portalservice.app.responses.PageResponse;
+import com.jun.portalservice.domain.entities.mongo.Account;
+import com.jun.portalservice.domain.entities.mongo.Voucher;
+import com.jun.portalservice.domain.entities.mongo.VoucherAccount;
+import com.jun.portalservice.domain.entities.types.GiveVoucherType;
 import com.jun.portalservice.domain.entities.types.VoucherState;
 import com.jun.portalservice.domain.exceptions.BadRequestException;
 import com.jun.portalservice.domain.exceptions.ResourceNotFoundException;
-import com.jun.portalservice.app.dtos.VoucherDTO;
-import com.jun.portalservice.app.responses.PageResponse;
-import com.jun.portalservice.domain.entities.mongo.Voucher;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.core.query.Criteria;
@@ -15,6 +19,7 @@ import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -22,29 +27,25 @@ import java.util.List;
 public class VoucherService extends BaseService {
 
   @Transactional(isolation = Isolation.SERIALIZABLE)
-  public Voucher create(VoucherDTO dto, String userId) throws Exception {
-    List<Voucher> voucherList = new ArrayList<>();
-
-    for (int i = 0; i < 30; i++) {
-      if (dto == null) {
-        throw new BadRequestException("RequestBody is null!");
-      }
-      if (userId == null) {
-        throw new BadRequestException("UserId is null!");
-      }
-
-      Voucher voucher = new Voucher();
-      voucher.from(dto);
-      voucher.setCreatedBy(userId);
-      voucher.setState(VoucherState.ACTIVE);
-      voucherList.add(voucher);
-      voucherStorage.save(voucher);
+  public Voucher create(VoucherDTO dto, Integer userId) throws Exception {
+    if (dto == null) {
+      throw new BadRequestException("RequestBody is null!");
     }
-    return voucherList.get(1);
+    if (userId == null) {
+      throw new BadRequestException("UserId is null!");
+    }
+
+    Voucher voucher = new Voucher();
+    voucher = modelMapper.toVoucher(dto);
+    voucher.setExpiryDate(
+        LocalDateTime.parse(dto.getDate(), DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm")));
+    voucher.setCreatedBy(userId);
+    voucher.setId((int) generateSequence(Voucher.SEQUENCE_NAME));
+    return voucherStorage.save(voucher);
   }
 
   @Transactional(isolation = Isolation.SERIALIZABLE)
-  public Voucher update(VoucherDTO dto, String voucherId, String userId) throws Exception {
+  public Voucher update(VoucherDTO dto, Integer voucherId, Integer userId) throws Exception {
     if (dto == null) {
       throw new BadRequestException("RequestBody is null!");
     }
@@ -60,50 +61,48 @@ public class VoucherService extends BaseService {
       throw new ResourceNotFoundException("Voucher " + voucherId + " is not found!");
     }
 
-    voucher.from(dto);
-    if (voucher.getState() != VoucherState.ACTIVE
-        && dto.getState() != null
-        && dto.getState() == VoucherState.ACTIVE) {
-      voucher.setActivatedAt(LocalDateTime.now());
-    }
+    voucher = modelMapper.toVoucher(dto);
+    voucher.setId(voucherId);
+
+    voucher.setExpiryDate(
+        LocalDateTime.parse(dto.getDate(), DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm")));
 
     voucher.setState(
         dto.getState() != null
             ? dto.getState()
             : voucher.getState() != null ? voucher.getState() : VoucherState.INACTIVE);
-    voucher.setUpdatedBy(userId);
 
     return voucherStorage.save(voucher);
   }
 
-  public Page<Voucher> listVoucher(String voucherId, String userId, Pageable pageable) {
+  public PageResponse<Voucher> listVoucher(
+      String voucherId, String code, String name, Pageable pageable) {
 
     List<Criteria> andConditions = new ArrayList<>();
-
+    andConditions.add(Criteria.where("id").ne(null));
     if (voucherId != null) {
       andConditions.add(Criteria.where("id").is(voucherId));
     }
-    if (userId != null) {
-      andConditions.add(Criteria.where("created_by").is(userId));
+    if (StringUtils.isNotEmpty(name)) {
+      andConditions.add(Criteria.where("name").regex(name, "i"));
+    }
+    if (StringUtils.isNotEmpty(code)) {
+      andConditions.add(Criteria.where("code").regex(code, "i"));
     }
 
-    if (andConditions == null || andConditions.size() == 0) {
-      andConditions.add(Criteria.where("id").ne(null));
-    }
     Query query = new Query();
     Criteria andCriteria = new Criteria();
 
-    query.addCriteria(
-        andCriteria.andOperator(andConditions.toArray(new Criteria[andConditions.size()])));
+    query.addCriteria(andCriteria.andOperator(andConditions.toArray(new Criteria[0])));
 
     Page<Voucher> voucherPage = voucherStorage.findAll(query, pageable);
 
     //    PageResponse pageResponse = new PageResponse(voucherPage,)
 
-    return voucherPage;
+    return PageResponse.createFrom(voucherPage);
   }
 
-  public Voucher findById(String voucherId) {
+  public Voucher findById(Integer voucherId) {
     if (voucherId == null) {
       throw new BadRequestException("VocherId must be not null!");
     }
@@ -128,13 +127,60 @@ public class VoucherService extends BaseService {
   }
 
   @Transactional(isolation = Isolation.SERIALIZABLE)
-  public boolean delete(String voucherId) {
+  public boolean delete(Integer voucherId) {
+
+    Voucher voucher = voucherStorage.findById(voucherId);
+
+    if (voucher == null) {
+      throw new ResourceNotFoundException("no vouhcer fount!");
+    }
+
+    if (voucher.getExpiryDate().isAfter(LocalDateTime.now())) {
+      throw new BadRequestException("Voucher is in expiry date");
+    }
+
     return voucherStorage.delete(voucherId);
   }
 
-  public PageResponse<Voucher> findAll(Pageable pageable) {
-    Page<Voucher> voucherPage = voucherStorage.findAll(pageable);
+  //  public static void main(String[] args) {
+  //    System.out.println(LocalDateTime.now());
+  //
+  //    LocalDateTime localDateTime = LocalDateTime.of(2021, 11, 15, 15, 18);
+  //    System.out.println(localDateTime.isAfter(LocalDateTime.now()));
+  //  }
 
-    return PageResponse.createFrom(voucherPage);
+  //  public PageResponse<Voucher> findAll(Pageable pageable) {
+  //    Page<Voucher> voucherPage = voucherStorage.findAll(pageable);
+  //
+  //    return PageResponse.createFrom(voucherPage);
+  //  }
+
+  public Boolean giveVoucher(GiveVoucherType type, List<Integer> accountIds, Integer voucherId) {
+
+    Voucher voucher = voucherStorage.findById(voucherId);
+    if (voucher == null) {
+      throw new ResourceNotFoundException("No voucher fount!");
+    }
+    List<Account> accounts = new ArrayList<>();
+
+    if (type == GiveVoucherType.ALL) {
+      accounts = accountRepository.findAll();
+    } else {
+      if (accountIds == null) {
+        throw new BadRequestException("List accountId null");
+      }
+      accounts = accountRepository.findAllById(accountIds);
+    }
+
+    List<VoucherAccount> voucherAccounts = new ArrayList<>();
+    for (Account account : accounts) {
+      VoucherAccount voucherAccount = new VoucherAccount(account.getId(), voucher);
+      voucherAccount.setIsUsed(false);
+      voucherAccounts.add(voucherAccount);
+
+      voucherStorage.deleteCache();
+    }
+    voucherAccountRepository.saveAll(voucherAccounts);
+    return true;
   }
 }
